fix(console): should not display unsaved alert on item deleted
